proj_810102198_810102151_810103109/requirements.txt

joblib
scikit-image

torch
torchvision
numpy
pandas
opencv-python
matplotlib
seaborn
scikit-learn
tqdm
python-Levenshtein